fix: this keyword, changed inelegant solution for service version
I've changed to logic for checking and setting the service version. Now it's redundant, but I couldn't wrap my head around what logic I should use to avoid using the same 2 lines of code when the service does not exists and when it should be updated because of the service version change.