v1.2.7
* feat(AC): Emulate HomeKit heat and cool thresholds

HomeKit expects separate heat and cool thresholds, but these ACs expose a
single target. This change emulates support for these thresholds and makes the
HomeKit’s HeaterCooler interface work properly on Auto mode.

To drive the device single TARGET_TEMPERATURE a new method
`setTargetTemperatureWithinThresholds` was added that has the following
behaviour:
  - COOL: target temperature is the cooling threshold
  - HEAT: target temperature is the heating threshold
  - AUTO: if the indoor temperature is below the heating threshold, set
    the target temperature to the heating threshold; if the indoor
    temperature is above the cooling threshold, set the target
    temperature to the cooling threshold; if the temperature is within
    the band set the target temperature to the current indoor
    temperature so the devices stays idle and save energy. For these
    auto adjustments prompt tone gets tempoerarily disabled.

The emulated thresholds are persisted in the accessory.context and
should survive homebridge restarts.

Changing the target temperature outside of HomeKit should adjust the
thresholds so they don't collapse and don't violate the invariant heating
threshold < cooling threshold.

The logic for determining the CurrentHeaterCoolerState was changed to
use the new thresholds.

Fixes: #134

* chore: reorder functions

* chore: added docs for auto mode

* chore: node and package version update

* fix: remove not needed screen display setter

---------

Co-authored-by: Patrik Kovalovszky <kovapatrik@gmail.com>